refactor : 쿼리문 수정 및 패키지 구조 변경
- 변경 사항 1: xml 파일 대문자 변경
- 변경 사항 2: 패키지 구조 변경
