+++ Pick/src/main/java/com/nob/pick/member/query/service/MemberService.java
@@ -1,4 +1,4 @@
-package com.nob.pick.member.query.application;
+package com.nob.pick.member.query.service;
 
 import java.util.HashMap;
 import java.util.List;
@@ -13,22 +13,22 @@ import com.nob.pick.member.query.dto.MemberProfilePageDTO;
 import com.nob.pick.member.query.dto.ProgrammingLanguageInfoDTO;
 import com.nob.pick.member.query.dto.Status;
 import com.nob.pick.member.query.dto.UserGrant;
-import com.nob.pick.member.query.infrastructure.repository.MemberRepository;
+import com.nob.pick.member.query.mapper.MemberMapper;
 
 @Service
 public class MemberService {
-	private final MemberRepository memberRepository;
+	private final MemberMapper memberMapper;
 
 	@Autowired
-	public MemberService(MemberRepository memberRepository) {
-		this.memberRepository = memberRepository;
+	public MemberService(MemberMapper memberMapper) {
+		this.memberMapper = memberMapper;
 	}
 
 	public String findEmail(String name, String phoneNumber) {
 		Map<String, Object> params = new HashMap<>();
 		params.put("name", name);
 		params.put("phoneNumber", phoneNumber);
-		return Optional.ofNullable(memberRepository.findEmailByNameAndPhone(params))
+		return Optional.ofNullable(memberMapper.findEmailByNameAndPhone(params))
 			.orElseThrow(() -> new IllegalArgumentException("이름과 전화번호로 회원을 찾을 수 없습니다."));
 	}
 
@@ -37,45 +37,45 @@ public class MemberService {
 		params.put("name", name);
 		params.put("phoneNumber", phoneNumber);
 		params.put("email", email);
-		return Optional.ofNullable(memberRepository.findPasswordByNamePhoneAndEmail(params))
+		return Optional.ofNullable(memberMapper.findPasswordByNamePhoneAndEmail(params))
 			.orElseThrow(() -> new IllegalArgumentException("이름과 전화번호, 이메일로 회원을 찾을 수 없습니다."));
 	}
 
 	public List<MemberDTO> findMemberInfo() {
-		return memberRepository.findAllMembers();
+		return memberMapper.findAllMembers();
 	}
 
 	public MemberDTO findMemberInfoById(int id) {
-		return Optional.ofNullable(memberRepository.findMemberById(id))
+		return Optional.ofNullable(memberMapper.findMemberById(id))
 			.orElseThrow(() -> new IllegalArgumentException("ID로 회원을 찾을 수 없습니다."));
 	}
 
 	public boolean existsEmail(String email) {
-		return memberRepository.existsByEmail(email);
+		return memberMapper.existsByEmail(email);
 	}
 
 	public boolean existsPhoneNumber(String phoneNumber) {
-		return memberRepository.existsByPhoneNumber(phoneNumber);
+		return memberMapper.existsByPhoneNumber(phoneNumber);
 	}
 
 	public Status findMemberStatus(int id) {
-		return memberRepository.findMemberStatusById(id);
+		return memberMapper.findMemberStatusById(id);
 	}
 
 	public UserGrant findUserGrant(int id) {
-		MemberDTO member = memberRepository.findUserGrantById(id);
+		MemberDTO member = memberMapper.findUserGrantById(id);
 		return member != null ? member.getUserGrant() : null;
 	}
 
 	public MemberProfilePageDTO findProfilePageByMemberId(int memberId) {
-		return memberRepository.findProfilePageByMemberId(memberId);
+		return memberMapper.findProfilePageByMemberId(memberId);
 	}
 
 	public List<ProgrammingLanguageInfoDTO> findProgrammingLanguagesByProfilePageId(int profilePageId) {
-		return memberRepository.findProgrammingLanguagesByProfilePageId(profilePageId);
+		return memberMapper.findProgrammingLanguagesByProfilePageId(profilePageId);
 	}
 
 	public List<ProgrammingLanguageInfoDTO> findActiveProgrammingLanguages() {
-		return memberRepository.findActiveProgrammingLanguages();
+		return memberMapper.findActiveProgrammingLanguages();
 	}
 }
