+++ Pick/src/main/resources/mapper/MemberMapper.xml
@@ -2,7 +2,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.nob.pick.member.query.infrastructure.repository.MemberRepository">
+<mapper namespace="com.nob.pick.member.query.mapper.MemberMapper">
     <!-- 회원 관련 resultMap -->
     <resultMap id="memberDTOMap" type="com.nob.pick.member.query.dto.MemberDTO">
         <id property="id" column="id"/>
@@ -39,115 +39,116 @@
 
     <!-- 이름과 전화번호로 이메일 찾기 -->
     <select id="findEmailByNameAndPhone" parameterType="map" resultType="string">
-        SELECT email
-        FROM member
-        WHERE name = #{name}
-          AND phone_number = #{phoneNumber}
+        SELECT M.EMAIL
+        FROM MEMBER M
+        WHERE M.NAME = #{name}
+          AND M.PHONE_NUMBER = #{phoneNumber}
     </select>
 
     <!-- 이름, 전화번호, 이메일로 비밀번호 찾기 -->
     <select id="findPasswordByNamePhoneAndEmail" parameterType="map" resultType="string">
-        SELECT password
-        FROM member
-        WHERE name = #{name}
-          AND phone_number = #{phoneNumber}
-          AND email = #{email}
+        SELECT M.PASSWORD
+        FROM MEMBER M
+        WHERE M.NAME = #{name}
+          AND M.PHONE_NUMBER = #{phoneNumber}
+          AND M.EMAIL = #{email}
     </select>
 
     <!-- 모든 회원 조회 -->
     <select id="findAllMembers" resultType="com.nob.pick.member.query.dto.MemberDTO">
-        SELECT id,
-               name,
-               age,
-               ihidnum,
-               phone_number,
-               email,
-               password,
-               nickname,
-               status,
-               reported_count,
-               user_grant
-        FROM member
+        SELECT M.ID,
+               M.NAME,
+               M.AGE,
+               M.IHIDNUM,
+               M.PHONE_NUMBER,
+               M.EMAIL,
+               M.PASSWORD,
+               M.NICKNAME,
+               M.STATUS,
+               M.REPORTED_COUNT,
+               M.USER_GRANT
+        FROM MEMBER M
     </select>
 
     <!-- ID로 특정 회원 조회 -->
     <select id="findMemberById" parameterType="int" resultType="com.nob.pick.member.query.dto.MemberDTO">
-        SELECT id,
-               name,
-               age,
-               ihidnum,
-               phone_number,
-               email,
-               password,
-               nickname,
-               status,
-               reported_count,
-               user_grant
-        FROM member
-        WHERE id = #{id}
+        SELECT M.ID,
+               M.NAME,
+               M.AGE,
+               M.IHIDNUM,
+               M.PHONE_NUMBER,
+               M.EMAIL,
+               M.PASSWORD,
+               M.NICKNAME,
+               M.STATUS,
+               M.REPORTED_COUNT,
+               M.USER_GRANT
+        FROM MEMBER M
+        WHERE M.ID = #{id}
     </select>
 
     <!-- ID로 회원 상태 조회 -->
     <select id="findMemberStatusById" parameterType="int" resultType="com.nob.pick.member.query.dto.Status">
-        SELECT status
-        FROM member
-        WHERE id = #{id}
+        SELECT M.STATUS
+        FROM MEMBER M
+        WHERE M.ID = #{id}
     </select>
 
     <!-- 이메일 중복 여부 확인 -->
     <select id="existsByEmail" parameterType="string" resultType="boolean">
         SELECT EXISTS(SELECT 1
-                      FROM member
-                      WHERE email = #{email})
+                      FROM MEMBER M
+                      WHERE M.EMAIL = #{email})
     </select>
 
     <!-- 전화번호 중복 여부 확인 -->
     <select id="existsByPhoneNumber" parameterType="string" resultType="boolean">
         SELECT EXISTS(SELECT 1
-                      FROM member
-                      WHERE phone_number = #{phoneNumber})
+                      FROM MEMBER M
+                      WHERE M.PHONE_NUMBER = #{phoneNumber})
     </select>
 
-    <!-- 유저 권한 확인  -->
+    <!-- 유저 권한 확인 -->
     <select id="findUserGrantById" parameterType="int" resultType="com.nob.pick.member.query.dto.MemberDTO">
-        SELECT user_grant
-        FROM member
-        WHERE id = #{id}
+        SELECT M.USER_GRANT
+        FROM MEMBER M
+        WHERE M.ID = #{id}
     </select>
 
+    <!-- 회원 프로필 페이지 조회 -->
     <select id="findProfilePageByMemberId" parameterType="int" resultMap="memberProfilePageDTOMap">
-        SELECT mp.id,
-               mp.level,
-               mp.introduce,
-               mp.preferred_area,
-               mp.score,
-               mp.image_path,
-               mp.member_id,
-               m.name AS member_name
-        FROM member_profile_page mp
-                 LEFT JOIN member m ON mp.member_id = m.id
-        WHERE mp.member_id = #{memberId}
+        SELECT MP.ID,
+               MP.LEVEL,
+               MP.INTRODUCE,
+               MP.PREFERRED_AREA,
+               MP.SCORE,
+               MP.IMAGE_PATH,
+               MP.MEMBER_ID,
+               M.NAME AS MEMBER_NAME
+        FROM MEMBER_PROFILE_PAGE MP
+                 LEFT JOIN MEMBER M ON MP.MEMBER_ID = M.ID
+        WHERE MP.MEMBER_ID = #{memberId}
     </select>
 
     <!-- 회원별 사용 언어 조회 -->
     <select id="findProgrammingLanguagesByProfilePageId" parameterType="int" resultMap="programmingLanguageInfoMap">
-        SELECT mpl.id,
-               mpl.programming_language_id,
-               pl.language
-        FROM member_profile_page mp
-                 JOIN member_programming_language mpl ON mp.id = mpl.member_profile_page_id
-                 JOIN programming_language pl ON mpl.programming_language_id = pl.id
-        WHERE mp.id = #{profilePageId}
-          AND pl.is_deleted = 'N'
+        SELECT MPL.ID,
+               MPL.PROGRAMMING_LANGUAGE_ID,
+               PL.LANGUAGE
+        FROM MEMBER_PROFILE_PAGE MP
+                 JOIN MEMBER_PROGRAMMING_LANGUAGE MPL ON MP.ID = MPL.MEMBER_PROFILE_PAGE_ID
+                 JOIN PROGRAMMING_LANGUAGE PL ON MPL.PROGRAMMING_LANGUAGE_ID = PL.ID
+        WHERE MP.ID = #{profilePageId}
+          AND PL.IS_DELETED = 'N'
     </select>
 
     <!-- 프로그래밍 언어 목록 조회 (N인 경우만) -->
     <select id="findActiveProgrammingLanguages" resultMap="programmingLanguageInfoMap">
-        SELECT id,
-               NULL AS programming_language_id,
-               language
-        FROM programming_language
-        WHERE is_deleted = 'N'
+        SELECT PL.ID,
+               NULL AS PROGRAMMING_LANGUAGE_ID,
+               PL.LANGUAGE
+        FROM PROGRAMMING_LANGUAGE PL
+        WHERE PL.IS_DELETED = 'N'
     </select>
 
 </mapper>
